refactor: add platform package in pkg/oci
