+++ cmd/cosign/cli/download/sbom.go
@@ -23,10 +23,9 @@ import (
 	"os"
 
 	"github.com/google/go-containerregistry/pkg/name"
-	v1 "github.com/google/go-containerregistry/pkg/v1"
-	"github.com/sigstore/cosign/v2/cmd/cosign/cli/common"
 	"github.com/sigstore/cosign/v2/cmd/cosign/cli/options"
 	"github.com/sigstore/cosign/v2/pkg/oci"
+	"github.com/sigstore/cosign/v2/pkg/oci/platform"
 	ociremote "github.com/sigstore/cosign/v2/pkg/oci/remote"
 )
 
@@ -49,43 +48,12 @@ func SBOMCmd(
 		return nil, err
 	}
 
+	se, err = platform.SignedEntityForPlatform(se, dnOpts.Platform)
+	if err != nil {
+		return nil, err
+	}
+
 	idx, isIndex := se.(oci.SignedImageIndex)
-
-	// We only allow --platform on multiarch indexes
-	if dnOpts.Platform != "" && !isIndex {
-		return nil, fmt.Errorf("specified reference is not a multiarch image")
-	}
-
-	if dnOpts.Platform != "" && isIndex {
-		targetPlatform, err := v1.ParsePlatform(dnOpts.Platform)
-		if err != nil {
-			return nil, fmt.Errorf("parsing platform: %w", err)
-		}
-		platforms, err := common.GetIndexPlatforms(idx)
-		if err != nil {
-			return nil, fmt.Errorf("getting available platforms: %w", err)
-		}
-
-		platforms = common.MatchPlatform(targetPlatform, platforms)
-		if len(platforms) == 0 {
-			return nil, fmt.Errorf("unable to find an SBOM for %s", targetPlatform.String())
-		}
-		if len(platforms) > 1 {
-			return nil, fmt.Errorf(
-				"platform spec matches more than one image architecture: %s",
-				platforms.String(),
-			)
-		}
-
-		nse, err := idx.SignedImage(platforms[0].Hash)
-		if err != nil {
-			return nil, fmt.Errorf("searching for %s image: %w", platforms[0].Hash.String(), err)
-		}
-		if nse == nil {
-			return nil, fmt.Errorf("unable to find image %s", platforms[0].Hash.String())
-		}
-		se = nse
-	}
 
 	file, err := se.Attachment("sbom")
 	if errors.Is(err, ociremote.ErrImageNotFound) {
@@ -93,7 +61,7 @@ func SBOMCmd(
 			return nil, errors.New("no sbom attached to reference")
 		}
 		// Help the user with the available architectures
-		pl, err := common.GetIndexPlatforms(idx)
+		pl, err := platform.GetIndexPlatforms(idx)
 		if len(pl) > 0 && err == nil {
 			fmt.Fprintf(
 				os.Stderr,
